refactor diff
