+++ lapce-app/Cargo.toml
@@ -48,8 +48,8 @@ sled = "0.34.7"
 bytemuck = "1.8.0"
 tokio = { version = "1.21", features = ["full"] }
 futures = "0.3.26"
-# floem = { git = "https://github.com/lapce/floem", rev = "7bcd3725290765f9d69374b883a13283646daac8" }
-floem = { path = "../../workspaces/floem" }
+floem = { git = "https://github.com/lapce/floem", rev = "17ffb5577d9519d04b1871b01e60da70f0765b11" }
+# floem = { path = "../../workspaces/floem" }
 config = { version = "0.13.2", default-features = false, features = ["toml"] }
 structdesc = { git = "https://github.com/lapce/structdesc" }
